+++ onnxscript/function_libs/torch_lib/graph_building_test.py
@@ -11,7 +11,7 @@ import onnxscript.testing
 from onnxscript import FLOAT, evaluator
 from onnxscript import opset17 as op
 from onnxscript._internal import version_utils
-from onnxscript.function_libs.torch_aten import graph_building, ops
+from onnxscript.function_libs.torch_lib import graph_building, ops
 
 
 @unittest.skipIf(version_utils.torch_older_than("2.0"), "torchscript in 1.13 not supported")
