+++ onnxscript/tests/function_libs/torch_lib/ops_correctness_test.py
@@ -59,11 +59,11 @@ from torch.utils import _pytree as pytree
 import onnxscript
 import onnxscript.evaluator
 from onnxscript._internal import version_utils
-from onnxscript.function_libs.torch_aten import graph_building
-from onnxscript.function_libs.torch_aten.ops import core as core_ops
-from onnxscript.function_libs.torch_aten.ops import nn as nn_ops
-from onnxscript.function_libs.torch_aten.ops import special as special_ops
-from onnxscript.tests.function_libs.torch_aten import extra_opinfo
+from onnxscript.function_libs.torch_lib import graph_building
+from onnxscript.function_libs.torch_lib.ops import core as core_ops
+from onnxscript.function_libs.torch_lib.ops import nn as nn_ops
+from onnxscript.function_libs.torch_lib.ops import special as special_ops
+from onnxscript.tests.function_libs.torch_lib import extra_opinfo
 
 T = TypeVar("T")
